refactor : exception 모듈 구분
